SOLR-14381: Grouping to use long instead of int for matches and ngroups to avoid overflows

+++ solr/core/src/java/org/apache/solr/request/SimpleFacets.java
@@ -343,7 +343,7 @@ public class SimpleFacets {
         .add(mainQueryFilter, Occur.FILTER)
         .build();
     searcher.search(filteredFacetQuery, collector);
-    return collector.getGroupCount();
+    return (int) collector.getGroupCount();
   }
 
   enum FacetMethod {
